+++ packages/components/src/DataList/DataList.test.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable max-statements */
 import { act, fireEvent, render, screen, within } from "@testing-library/react";
-import React from "react";
+import React, { ReactElement } from "react";
 import { configMocks, mockIntersectionObserver } from "jsdom-testing-mocks";
 import userEvent from "@testing-library/user-event";
 import { Banner } from "@jobber/components/Banner";
@@ -10,6 +10,7 @@ import {
   Breakpoints,
   DATA_LIST_FILTERING_SPINNER_TEST_ID,
   DATA_LIST_LOADING_MORE_SPINNER_TEST_ID,
+  DATA_LIST_STICKY_HEADER_TEST_ID,
   DATA_LOAD_MORE_TEST_ID,
   EMPTY_FILTER_RESULTS_MESSAGE,
 } from "./DataList.const";
@@ -166,10 +167,12 @@ describe("DataList", () => {
   describe("Layout", () => {
     const layoutWrapper = "layout-wrapper";
     const layoutItem = "layout-item";
+    const mockOnSearch = jest.fn();
 
     beforeEach(() => {
       render(
         <DataList data={mockData} headers={mockHeaders}>
+          <DataList.Search onSearch={mockOnSearch} />
           <DataList.Layout>
             {(item: DataListItemType<typeof mockData>) => (
               <div data-testid={layoutWrapper}>
@@ -313,6 +316,7 @@ describe("DataList", () => {
       }) => {
         const layout1Wrapper = "layout1-wrapper";
         const layout2Wrapper = "layout2-wrapper";
+        const mockOnSearch = jest.fn();
         setUpMediaQueries(mockedQueries);
         render(
           <DataList
@@ -320,6 +324,7 @@ describe("DataList", () => {
             headers={mockHeaders}
             headerVisibility={headerVisibility}
           >
+            <DataList.Search onSearch={mockOnSearch} />
             <DataList.Layout size={layoutSize1}>
               {(item: DataListItemType<typeof mockData>) => (
                 <div data-testid={layout1Wrapper}>
@@ -346,6 +351,8 @@ describe("DataList", () => {
   });
 
   describe("Header", () => {
+    const mockOnSearch = jest.fn();
+
     function renderLayout(
       headerVisibility?: DataListProps<
         (typeof mockData)[0]
@@ -359,6 +366,7 @@ describe("DataList", () => {
           headerVisibility={headerVisibility}
           sorting={sorting}
         >
+          <DataList.Search onSearch={mockOnSearch} />
           <DataList.Layout>
             {(item: DataListItemType<typeof mockData>) => (
               <div>{item.name}</div>
@@ -388,7 +396,46 @@ describe("DataList", () => {
     });
   });
 
+  describe("DataListStickyHeader", () => {
+    const mockOnSearch = jest.fn();
+
+    const renderDataListWithLayout = (children: ReactElement) => {
+      render(
+        <DataList data={mockData} headers={{}}>
+          {children}
+        </DataList>,
+      );
+    };
+
+    it("should render the Sticky Header if DataList.Search is provided", () => {
+      renderDataListWithLayout(<DataList.Search onSearch={mockOnSearch} />);
+      expect(
+        screen.queryByTestId(DATA_LIST_STICKY_HEADER_TEST_ID),
+      ).toBeInTheDocument();
+    });
+
+    it("should render the Sticky Header if DataList.Filters is provided", () => {
+      renderDataListWithLayout(
+        <DataList.Filters>
+          <div>Filters</div>
+        </DataList.Filters>,
+      );
+      expect(
+        screen.queryByTestId(DATA_LIST_STICKY_HEADER_TEST_ID),
+      ).toBeInTheDocument();
+    });
+
+    it("should not render Sticky Header when DataList.Search and DataList.Filters are not provided", () => {
+      renderDataListWithLayout(<></>);
+      expect(
+        screen.queryByTestId(DATA_LIST_STICKY_HEADER_TEST_ID),
+      ).not.toBeInTheDocument();
+    });
+  });
+
   describe("Sorting", () => {
+    const mockOnSearch = jest.fn();
+
     function MockSortingLayout({
       sorting,
     }: {
@@ -396,6 +443,7 @@ describe("DataList", () => {
     }) {
       return (
         <DataList data={mockData} headers={mockHeaders} sorting={sorting}>
+          <DataList.Search onSearch={mockOnSearch} />
           <DataList.Layout>
             {(item: DataListItemType<typeof mockData>) => (
               <div>{item.name}</div>
@@ -566,6 +614,7 @@ describe("DataList", () => {
     it("should show the StatusBar when it's provided", () => {
       const bannerText =
         "Something went wrong. Refresh or check your internet connection.";
+      const mockOnSearch = jest.fn();
       render(
         <DataList
           data={Array.from({ length: MAX_DATA_COUNT + 1 }, (_, id) => ({
@@ -573,6 +622,7 @@ describe("DataList", () => {
           }))}
           headers={{ id: "ID" }}
         >
+          <DataList.Search onSearch={mockOnSearch} />
           <DataList.StatusBar>
             <Banner type="error" icon="alert">
               {bannerText}